added Queue and Graph implementation
The graph implementation can be shown using a network button in the app that will pop up an adjacency list that will show the connection between friends.
The Queue implementation was used to limit the display of recently played games in the profile to only four at a time. Old entries will be removed.