ADDED Stack Implementation and LinkedList ADT
Stack Implementation was added in the JashanProject because we changed the question into a combination of operators. in order to check the value, stack ADT was really needed for the infix to postfix method. 
LinkedList was for the Billionaire.java since LinkedList could do the things Queue with much ease. This makes the LinkedList much more needed in the application.